add click tracker hoc and context
