+++ src/Components/RelatedItems/comp/Carousel.jsx
@@ -1,5 +1,6 @@
 import React from 'react';// Bring React in to build a component.
 import {useState, useEffect} from 'react';
+import ClickTracker from '../../../ClickTracker.jsx';
 
 
 var Carousel = ({items}) => {
