+++ src/ElementContext.js
@@ -1,0 +1,5 @@
+import React, {createContext} from 'react';
+
+const ElementContext = createContext();
+
+export default ElementContext;
