+++ src/ModuleContext.js
@@ -1,0 +1,5 @@
+import React, {createContext} from 'react';
+
+const ModuleContext = createContext();
+
+export default ModuleContext;
